+++ Source/DafnyCore/Compilers/GoLang/Compiler-go.cs
