+++ Source/DafnyCore/Compilers/SinglePassCompiler.cs
@@ -16,21 +16,6 @@ using Microsoft.BaseTypes;
 
 
 namespace Microsoft.Dafny.Compilers {
-  internal class InternalCompilersPluginConfiguration : Plugins.PluginConfiguration {
-    public static readonly InternalCompilersPluginConfiguration Singleton = new();
-
-    public override Plugins.Compiler[] GetCompilers() {
-      return new Plugins.Compiler[] {
-        new Compilers.CsharpBackend(),
-        new Compilers.JavaScriptCompiler(),
-        new Compilers.GoCompiler(),
-        new Compilers.JavaCompiler(),
-        new Compilers.PythonCompiler(),
-        new Compilers.CppCompilerBackend()
-      };
-    }
-  }
-
   public abstract class SinglePassCompiler {
     
     /// <summary>
@@ -60,7 +45,7 @@ namespace Microsoft.Dafny.Compilers {
     protected virtual string ModuleSeparator { get => "."; }
     protected virtual string ClassAccessor { get => "."; }
     
-    protected ErrorReporter? Reporter;
+    protected ErrorReporter Reporter;
 
     Stack<ConcreteSyntaxTree> copyInstrWriters = new Stack<ConcreteSyntaxTree>(); // a buffer that stores copy instructions generated by letExpr that uses out param.
     protected TopLevelDeclWithMembers thisContext;  // non-null when type members are being translated
@@ -83,7 +68,7 @@ namespace Microsoft.Dafny.Compilers {
       return n;
     }
 
-    public CoverageInstrumenter Coverage;
+    public readonly CoverageInstrumenter Coverage;
 
     protected SinglePassCompiler(ErrorReporter reporter) {
       Reporter = reporter;
@@ -1526,12 +1511,7 @@ namespace Microsoft.Dafny.Compilers {
       WriteFromStream(rd, wr.Append((new Verbatim())));
     }
 
-    protected void WriteFromFile(string inputFilename, TextWriter outputWriter) {
-      var rd = new StreamReader(new FileStream(inputFilename, FileMode.Open, FileAccess.Read));
-      WriteFromStream(rd, outputWriter);
-    }
-
-    protected void WriteFromStream(StreamReader rd, TextWriter outputWriter) {
+    public static void WriteFromStream(StreamReader rd, TextWriter outputWriter) {
       while (true) {
         string s = rd.ReadLine();
         if (s == null) {
