+++ Source/DafnyCore/DafnyOptions.cs
@@ -201,7 +201,7 @@ NoGhost - disable printing of functions, ghost methods, and proof
       Prune = true;
       NormalizeNames = true;
       EmitDebugInformation = false;
-      Compiler = new CsharpBackend();
+      Backend = new CsharpBackend();
     }
 
     public override string VersionNumber {
@@ -263,7 +263,7 @@ NoGhost - disable printing of functions, ghost methods, and proof
     public List<string> MainArgs = new List<string>();
 
     public string CompilerName;
-    public Compiler Compiler;
+    public IExecutableBackend Backend;
     public bool CompileVerbose = true;
     public bool EnforcePrintEffects = false;
     public string DafnyPrintCompiledFile = null;
@@ -811,7 +811,7 @@ NoGhost - disable printing of functions, ghost methods, and proof
     public void ApplyDefaultOptionsWithoutSettingsDefault() {
       base.ApplyDefaultOptions();
 
-      Compiler ??= new CsharpBackend();
+      Backend ??= new CsharpBackend();
 
       // expand macros in filenames, now that LogPrefix is fully determined
